+++ src/main/java/es/chollotek/beans/LineaPedido.java
@@ -8,10 +8,10 @@ import java.io.Serializable;
  */
 public class LineaPedido implements Serializable {
 
-    private short idlinea;
-    private short idpedido;
-    private short idproducto;
-    private short cantidad;
+    private int idlinea;
+    private int idpedido;
+    private int idproducto;
+    private int cantidad;
 
     public LineaPedido() {
     }
@@ -22,35 +22,35 @@ public class LineaPedido implements Serializable {
 //    }
 
     //Getters y Setters
-    public short getIdlinea() {
+    public int getIdlinea() {
         return idlinea;
     }
 
-    public void setIdlinea(short idlinea) {
+    public void setIdlinea(int idlinea) {
         this.idlinea = idlinea;
     }
 
-    public short getIdpedido() {
+    public int getIdpedido() {
         return idpedido;
     }
 
-    public void setIdpedido(short idpedido) {
+    public void setIdpedido(int idpedido) {
         this.idpedido = idpedido;
     }
 
-    public short getIdproducto() {
+    public int getIdproducto() {
         return idproducto;
     }
 
-    public void setIdproducto(short idproducto) {
+    public void setIdproducto(int idproducto) {
         this.idproducto = idproducto;
     }
 
-    public short getCantidad() {
+    public int getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(short cantidad) {
+    public void setCantidad(int cantidad) {
         this.cantidad = cantidad;
     }
